Refactor & fix
