--- src/worldofzuul/Pest.java
@@ -1,4 +1,0 @@
-package worldofzuul;
-
-public class Pest {
-}
