+++ src/worldofzuul/item/Disease.java
@@ -1,4 +1,4 @@
-package worldofzuul;
+package worldofzuul.item;
 
 public class Disease {
 }
