+++ src/worldofzuul/item/Pest.java
@@ -1,0 +1,4 @@
+package worldofzuul.item;
+
+public class Pest {
+}
